Create FIPA
A copy of Mass General Laws chapter 66a, the Fair Information Practices Act of the Commonwealth of Massachusetts.  Notable due to it's comprehensive scope as applied to personal data held by public sector agencies of the Commonwealth and the fact that is contains reporting requirements as well as enforcement with judicial remedies to correct violations of the rules.